[refactor]: exception 패키지 통일
