+++ src/main/java/com/sparta/sunday/domain/common/exception/EntityNotFoundException.java
@@ -1,4 +1,4 @@
-package com.sparta.sunday.exception;
+package com.sparta.sunday.domain.common.exception;
 
 public class EntityNotFoundException extends RuntimeException {
     public EntityNotFoundException(String message) {
