+++ src/main/java/com/sparta/sunday/domain/common/exception/UnAuthorizedException.java
@@ -1,4 +1,4 @@
-package com.sparta.sunday.exception;
+package com.sparta.sunday.domain.common.exception;
 
 public class UnAuthorizedException extends RuntimeException{
     public UnAuthorizedException(String message) {
